Fixes to namespace use in Tests project.
