+++ test/Altinn.Notifications.Tests/Mocks/EmailServiceMock.cs
@@ -1,10 +1,7 @@
-﻿using Altinn.Notifications.Core;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading;
+﻿using System.Threading;
 using System.Threading.Tasks;
+
+using Altinn.Notifications.Core;
 
 namespace Altinn.Notifications.Tests.Mocks
 {
