+++ test/Altinn.Notifications.Tests/Mocks/NotificationRepositoryMock.cs
@@ -1,10 +1,8 @@
-﻿using Altinn.Notifications.Core;
-using Altinn.Notifications.Core.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+
+using Altinn.Notifications.Core;
+using Altinn.Notifications.Core.Models;
 
 namespace Altinn.Notifications.Tests.Mocks
 {
